fix: warnings in shell and userland programs

+++ Userland/SampleCodeModule/include/programs/testsync.h
@@ -1,3 +1,3 @@
 #include <stdint.h>
 
-uint64_t test_sync(int argc, char *argv[]);
+int test_sync(int argc, char *argv[]);
